Move utils to internal package
